+++ srcs/Commands/CommandMode.cpp
@@ -33,7 +33,7 @@ int CommandMode::exec(Client *c)
             return c->sendMessage(RPL_CHANNELMODEIS, modesToString(chan->getMode()));
         if (c != chan->getCreator() || c->isOper())
             return c->sendMessage(ERR_CHANOPRIVSNEEDED, "Invalid privileges on the channel.");
-        for (int i = 1; i < argv[1].size(); i++)
+        for (size_t i = 1; i < argv[1].size(); i++)
         {
             if (argv[1][i] == 'i' || argv[1][i] == 'o' ||argv[1][i] == 'O' ||argv[1][i] == 'r' || argv[1][i] == 'w')
             {
@@ -52,7 +52,7 @@ int CommandMode::exec(Client *c)
             return c->sendMessage(ERR_USERSDONTMATCH, "Invalid user.");
         if (argv.size() == 1)
             return c->sendMessage(RPL_UMODEIS, modesToString(c->getMode()));
-        for (int i = 1; i < argv[1].size(); i++)
+        for (size_t i = 1; i < argv[1].size(); i++)
         {
             if (argv[1][i] == 'b' || argv[1][i] == 'e' ||argv[1][i] == 'l' ||argv[1][i] == 'i' || argv[1][i] == 'I' || argv[1][i] == 'k' || argv[1][i] == 'm' || argv[1][i] == 's')
             {
